完成 后台系统 系统管理模块的功能测试

+++ spmall-manager/src/main/java/com/project/ychengspmall/manager/helper/MenuHelper.java
@@ -5,7 +5,11 @@ import com.project.ychengspmall.model.entity.system.SysMenu;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * 构建菜单树的工具类helper
+ */
 public class MenuHelper {
+
     /**
      * 使用递归方法建菜单
      * @param sysMenuList
@@ -14,7 +18,8 @@ public class MenuHelper {
     public static List<SysMenu> buildTree(List<SysMenu> sysMenuList) {
         List<SysMenu> trees = new ArrayList<>();
         for (SysMenu sysMenu : sysMenuList) {
-            if (sysMenu.getParentId().longValue() == 0) {
+            //判断如果是最顶级顶点（parentID==0)，则寻找他的子节点
+            if (sysMenu.getParentId() == 0) {
                 trees.add(findChildren(sysMenu, sysMenuList));
             }
         }
@@ -27,6 +32,7 @@ public class MenuHelper {
      * @return
      */
     public static SysMenu findChildren(SysMenu sysMenu, List<SysMenu> treeNodes) {
+        //这里的sysMenu是父级id，只需要获取全部的TreeNodes中parentId为sysMenu的节点
         sysMenu.setChildren(new ArrayList<>());
         for (SysMenu it : treeNodes) {
             if (sysMenu.getId().longValue() == it.getParentId().longValue()) {
